FIx Club route
Added logo uploading
Fixed route
